+++ backend/api-application/src/main/java/com/homeloan/project/model/LoanApplication.java
@@ -18,9 +18,9 @@ public class LoanApplication {
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     private int id;
     private String address;
-    public double loan_amount;
+    public float loan_amount;
     public int tenure;
-    public double monthly_salary;
+    public float monthly_salary;
     private String status;
     private String saving_account_number;
 }
